move findNotExistingName to moving as it belongs there
